Moving diffs to fixture
